refactor:#36 이미지 업로드기능 리팩토링 (프로필 업로드 기능과 통합)
